server/requirements.txt
bcrypt
cryptography
flask
flask-bcrypt
flask-migrate
flask-sqlalchemy
gunicorn
marshmallow
PyMySQL
python-dotenv
pdfkit
flask-jwt-extended
flask-cors
flask-swagger-ui